Refactors GooglePlace into a single module
