+++ services/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     apt 'com.google.dagger:dagger-compiler:2.1'
     compile 'com.jakewharton.timber:timber:4.1.1'
     compile 'com.annimon:stream:1.0.6'
+    compile project(':domain')
 }
 
 // Using lambda functions
