+++ services/src/main/java/com/lyft/cityguide/services/google/place/GooglePlaceServiceFactory.java
@@ -1,11 +1,13 @@
 package com.lyft.cityguide.services.google.place;
 
-import com.lyft.cityguide.services.google.place.jobs.SearchPlacesJob;
+import com.google.gson.GsonBuilder;
 
 import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
+import retrofit.RestAdapter;
+import retrofit.converter.GsonConverter;
 
 /**
  * GooglePlaceServiceFactory
@@ -13,6 +15,34 @@ import dagger.Provides;
  */
 @Module
 public class GooglePlaceServiceFactory {
+
+    @Provides
+    IGooglePlaceAPI provideJSONAPI(ApplicationConfiguration configuration, SearchOutcomeGooglePlaceDTOSerializer serializer) {
+        return new RestAdapter.Builder()
+            .setEndpoint(configuration.GOOGLE_PLACE_API_ENDPOINT)
+            .setConverter(
+                new GsonConverter(
+                    new GsonBuilder()
+                        .registerTypeAdapter(SearchOutcomeGooglePlaceDTO.class, serializer)
+                        .create()
+                )
+            )
+            .build()
+            .create(IGooglePlaceAPI.class);
+    }
+
+    @Provides
+    @Singleton
+    SearchOutcomeGooglePlaceDTOSerializer provideSerializer() {
+        return new SearchOutcomeGooglePlaceDTOSerializer();
+    }
+
+    @Provides
+    @Singleton
+    SearchPlacesJob provideJob(SecretApplicationConfiguration configuration, IGooglePlaceAPI api, PointOfInterestBLLDTOSerializer serializer) {
+        return new SearchPlacesJob(configuration, api, serializer);
+    }
+
     @Provides
     @Singleton
     public IGooglePlaceService provideService(SearchPlacesJob searchPlacesJob) {
