+++ services/src/main/java/com/lyft/cityguide/services/google/place/SearchOutcomeGooglePlaceDTOSerializer.java
@@ -1,4 +1,4 @@
-package com.lyft.cityguide.services.google.place.api;
+package com.lyft.cityguide.services.google.place;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonDeserializationContext;
@@ -6,8 +6,6 @@ import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
-import com.lyft.cityguide.services.google.place.dto.PlaceGooglePlaceDTO;
-import com.lyft.cityguide.services.google.place.dto.SearchOutcomeGooglePlaceDTO;
 
 import java.lang.reflect.Type;
